Chainlit_FastAPI/requirements.txt
chainlit
cohere
openai
tiktoken
python-dotenv
pydantic